requirements.txt
Django==2.2
djangorestframework==3.12.2
pytz==2020.4
sqlparse==0.4.1
psycopg2>=2.8
Pillow>= 8.0
django-extensions==3.0.9
django-oauth-toolkit==1.3.3
pytest==6.2.0
pytest-django==4.1.0